Complete bookings migration
- Switch all code to use officeIds and new bookings table.
- Require officeId in config.
