+++ server/app.ts
@@ -46,7 +46,7 @@ export const configureApp = (config: Config) => {
   app.get('/api/config', (req, res, next) => {
     try {
       const clientConfig = {
-        version: '1.2.0',
+        version: '2.0.0',
         showTestBanner: config.showTestBanner,
         auth:
           config.authConfig.type === 'cognito'
